gcp/cloud-functions/twitter-ingestion/requirements.txt
functions-framework==3.5.0
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.10.0
tweepy==4.14.0